+++ CHANGELOG.rst
@@ -8,6 +8,8 @@ rpclib-2.8.0-beta
  * Make DateTime string format customizable.
  * Implement the server transport that exposes Application as a
    ``twisted.web.resource.Resource`` child.
+ * Remove Deprecated XMLAttribute and XMLAttributeRef. Use ``XmlAttribute``
+   and ``XmlAttributeRef`` instead.
 
 rpclib-2.7.0-beta
 -----------------